support publishing only pages needing updates
The following attempts to bring support for publishing pages which only
need to be updated. Pages are now published with a properly value of the
hash value of a page's contents used by this extension. When a future
publish event occurs, this hash is used to compare against newer pages.
If the hashes match, no new publish attempt will be made for the page.

Signed-off-by: James Knight <james.d.knight@live.com>