doc: improve look of code blocks inside admonitions
Code blocks inside admonitions appear to be wrapped with highlight divs
in recent Sphinx builds. To help provide a margin tweak for `pre` blocks
inside admonition, ensure the pre style override is applied for this
condition.

In additional, tweak the background color to be a bit lighter, to help
make the code block content a bit more distinguished inside an
admonition.

Signed-off-by: James Knight <james.d.knight@live.com>