translator: ensure v2 header flag is defined
Internally, the storage translator uses a `_v2_header_added` flag to
track header usage. An issue with this implementation is that if no
Confluence header is used, no page generation and no source links; the
flag value may not be defined when it may be used in some link checks.
To correct this, ensure the flag is initialized in the translation's
constructor.

Signed-off-by: James Knight <james.d.knight@live.com>
(cherry picked from commit 7cc8e4b6f75898dc4d539b22a734a74c390fd333)