pyproject.toml: add missing files in sdist
The most recent release (v2.1) generated a source distribution package
that did not include a license file. This was a result of an incomplete
switch from setuptools to flit. Correcting this by listing `LICENSE` as
a file to include in the source distribution (as well as other desired
files).

Signed-off-by: James Knight <james.d.knight@live.com>
(cherry picked from commit c76fcf8e1623909222f4075c7f84622ef8fa7dbd)