pyproject.toml: drop license key
The use of a desired SPDX license definition does not appear to be the
proper usage of the license text field (PEP-621 [1]). It also looks that
the use of this key as a table is to be deprecated (PEP-639 [2]). To
remove a possible incorrect license entry and not update the value to a
soon-to-be deprecated value, dropping its usage.

[1]: https://peps.python.org/pep-0621/
[2]: https://peps.python.org/pep-0639/

Signed-off-by: James Knight <james.d.knight@live.com>
(cherry picked from commit 2624604797840c62fd1062f6213fad51a984ee2b)