publisher: force publish on parent/config changes
As of v2.1, this extension no longer publishes pages when the content of
a page has not changed. This causes some undesired publications when
other configuration options change not related to content. To help
improve a users experience, we will allow force publishing for a page
on the following events:

- If the parent page to publish has changed.
- If there are new global labels missing from the existing page.
- If the configured full-width option has changed.
- If the editor type has changed.

Signed-off-by: James Knight <james.d.knight@live.com>