translator: support v1 limited-page-width on server/dc
When adding support for `confluence_full_width` with a v1 editor, the
styling used only worked gracefully on Confluence Cloud. For Confluence
Server/DC versions, the applied layout does not enforce any maximum
width on the contained content.

To improve this feature on Confluence Server/DC, we will use an
alternative formatting for page content. Instead, we will opt to using
a single div with a maximum width style applied. The maximum width
chosen is based on the observed default max-width for macros when using
the WYSIWYG editor.

Signed-off-by: James Knight <james.d.knight@live.com>