introduce permit-raw-html option
The following introduces the option `confluence_permit_raw_html` to
allow users to force-publish documents with raw HTML content in them.
While the use of this option will be unofficially supported, it is being
added solely since users keep requesting such a capability.

We unofficially supported some level of this using an undocumented
configuration option `confluence_adv_permit_raw_html`, but instead we
have renamed the option to `confluence_permit_raw_html` and have added
it into the configuration documentation (for users to easily find).

In addition, we have also tweaked the capability to also accept the name
of an HTML macro definition. This can allow users to place raw HTML in
a supported HTML macro (if enabled on their Confluence instance). While
this option is also still considered unsupported, having this provides
the most capability we can provide users who want to manage HTML-based
content as their own discretion.

Signed-off-by: James Knight <james.d.knight@live.com>