translator: always cast ac-params strings
When injecting ac:param tags into documentation, these values are
expected to be strings. Adjust the `_build_ac_param` method to always
cast the provided values as strings, to make it a bit more flexible for
callers to pass in non-string values but expect the values to be
automatically converted into strings.

Signed-off-by: James Knight <james.d.knight@live.com>