tox: adding ruff environment
Adds a tox environment and configuration for ruff checks. This is only
an initial introduction and not yet formally used/required for a formal
acceptance of implementation. There are several areas of improvements in
the implementation to be done, as well as possible adding some more
ignores for linting modes not supported (false-positives) and other
preference-related styles.

Over time, we can address warnings and then possible formally drop
flake8/pylint for just ruff.

Signed-off-by: James Knight <james.d.knight@live.com>