config: resolve relative ca-certificate
When this extension performs a configuration check for a CA certificate,
the check will pass if either the absolute path exists or that a
relative path based on the project's source directory exists. However,
after checking if the file exists, it passes the raw file into Requests
which may not be able to resolve a relative path. To prevent issues,
always determine the absolute path of the CA file/directory before
providing it to Requests.

Signed-off-by: James Knight <james.d.knight@live.com>