allow empty legacy pages set for new root content
When determining legacy pages when purging, the base identifier used is
either the parent container or a configured root page. When configured
to purge from root and the root page does not yet exist, this can give a
misrepresentation of what legacy pages are detected when a parent page
is also configured.

This commit changes this by force using the detected root page as the
cleanup root. And if the cleanup root does not exist, a `None` root
will result in an empty legacy pages set (which is expected).

Signed-off-by: James Knight <james.d.knight@live.com>