add http codes to error messages
While this extension tries to provide a descriptive error statement to
help users understand a failed scenario, it may be good to relay the
exact HTTP code for select events to provide users more feedback on the
state of things. This can be useful for advanced users who only want to
see the associated code with the instance they are interacting with.

Signed-off-by: James Knight <james.d.knight@live.com>