tests: always clone configuration to support decorator config overrides
When decorators are used to alter the test case's configuration, this
is applied to a base configuration prepared for the entire test class.
This could then result in some unit tests changing other unit test's
configuration, depending on how the unit test order is run. To prevent
unexpected configuration changes, always clone the base configuration
when running a prepare/build invoke.

Signed-off-by: James Knight <james.d.knight@live.com>