translator: nest v2 anchors inside paragraphs
For the v2 editor, we currently nest header anchors inside header blocks
to prevent the styling of the v2 editor for adding large spacing between
a leading anchor and the header.

When injecting anchors for paragraphs, the same spacing occurs. To avoid
this, we now also forward anchor creation to the start of a paragraph.

Signed-off-by: James Knight <james.d.knight@live.com>