support csv width hints in v2 editor
When using CSV tables with width hints, the v2 editor looks to ignore
the percentages applied on the column tags. However, if we provide an
explicit data width value on the table, percentages are taken into
consideration.

This commit tweaks the tables in v2 editor so that if column widths are
detected, we apply the default width typically given to a table in the
v2 editor to help allow column hints to work.

Signed-off-by: James Knight <james.d.knight@live.com>