publisher: relax various page data requirements
When publishing a page, relax the requirement to defining some keys for
a page (e.g. if a page has an explicit `full-width` value). This is to
help certain developer builders (e.g. sandbox mode) which may not
explicitly configure these values for a page, and do not care what
defaults are applied.

Signed-off-by: James Knight <james.d.knight@live.com>