validation: default to collapsed width for validation pages
Adjust the widths for validation pages to be collapsed width by default.
This makes all v1 pages render in the smaller-width environment,
providing a similar look to the v2 pages.

Although, we place an exception for the leading page based on how we
present the editor choice and custom widths for images used.

Signed-off-by: James Knight <james.d.knight@live.com>