validation: apply dummy target for base page images
When rendering images on the base page for both editors, users will
typically click on the images to jump to an editor selection, instead of
the header link. This causes the images to expand ~ where users then
unexpand then click on the header links.

While ideally users can click on these images to jump to the respective
editor document pages, Sphinx/docutils does not support processing
document links in the image's `target` option.

To deal with this scenario, we inject a custom reference link on the
current page. This then suppresses the zooming of the images and
"consuming" the click event.

Signed-off-by: James Knight <james.d.knight@live.com>