translator: correct v2 support tracking for documents
When processing individual documents through a translator, a v2 flag is
used to help easily indicate if the page is expected to be rendered in a
v1 or v2 editor. While the flag worked based on the global configuration
setting the editor, if any page-specific overrides are configured in the
meta directive, this could cause the flag to be incorrect set for a
document.

This commit moves the helper flag to be configured after processing the
metadata for a document, ensuring v2 is set considering both the global
configuration value and the page-specific override value.

Signed-off-by: James Knight <james.d.knight@live.com>