improve anchor links
The following provides a somewhat major rewrite on the creation of
anchors and how pages can link to these anchors. There has been some
limitations with linking to anchors when support for the v2 editor was
added. Ideally, pages will generate an `ac:link` macro and specify
anchors to target. However, testing showed that anchor links would not
always work with these macros since the macro expected anchors to be
prefixed with a page's name, which is not the case for v2 pages.

This commit attempts to fix these by building compatible anchor links on
pages no matter what editor version is configured for a page. This
extension now calculates expected targets for each document no matter
the editor (with an appropriate target expected for auto-generates
identifiers). When anchors are now added to a v2 page, we generate
additional anchors on a page that include a page-prefix, allowing any
`ac:link` macros to have an appropriate target to jump to (either the
old compatible link; or if sometime in the future, the new non-prefixed
link).

In addition, we also inject MyST slug-generated anchor points, to be
flexible for links created with the MyST-parser extension.

Note that there are still some limitations with anchors in the v2
editor. For example, anchors with parenthesis are silently dropped in
publish storage formatted documents. At this time, a support request has
been made with Atlassian to investigate this issue (JST-937603).

Signed-off-by: James Knight <james.d.knight@live.com>