builder: avoid unchanged page cache checks if we have legacy pages
Before attempting to cleanup any detected legacy pages from a target
Confluence instance, the builder will check to see if any unchanged
documents exist (pages that did not trigger an update and we not
removed from the internal legacy list), and remove them from the
legacy pages list if our cache detects it was already published.
However, this logic assumes we have populated a legacy pages list
ahead of time, which may not have happened if we are not configured
for any cleaning.

This commit adjusts the unchanged pages check to only occur if we have
populated a legacy pages list to check on.

Signed-off-by: James Knight <james.d.knight@live.com>