publisher: simplify initial connection check
The following refactors how the initial connection check is performed.
The original design introduced back in v0.6 used the "Get spaces" API
endpoint to validate access to a space. As the author of this change,
I could not recall the specifics of why it was chosen over "Get space"
(either an overlooked API call, possible issues observed when testing
initial REST support, etc.). At this time, there is no known reason
why a switch to the "Get space" API cannot be done. The new call appears
to work find for both Cloud and data server versions. Therefore, this
commit is making this change.

The switch to the explicit space API endpoint simplifies the publishers
logic. We no longer perform a related space key check (back in the day
where users may have added the Confluence name instead of space into a
legacy `confluence_space_name` option). This removes a special case
exception as well, generated for failed searches.

Note that the original and new endpoint switch being made is a
deprecated call on Confluence Cloud. This is to be addressed in the
future.

Signed-off-by: James Knight <james.d.knight@live.com>