builder: manage a pathlib output directory
Updating the implementation to manage a builder member variable
`out_dir`, which provides a path-like object instead of a string from
Sphinx's `outdir` variable. This commit also adjusts the implementation
that uses the output directory to take advantage of pathlib
capabilities.

This is part of an ongoing maintenance task to the implementation to use
newer capabilities of Python that are supported on the oldest interpreter
version this extension supports.

Signed-off-by: James Knight <james.d.knight@live.com>