enforce minimum sphinx v6.1 support
While the documentation of this extension indicates explicit support to
v6.1+, the Python project nor Sphinx version check never enforced this
requirement. This is partially due to some partial support for some
older versions (maybe back to Sphinx v5.x).

With a window open to perform some maintenance, adjusting the
configuration values to enforce the minimum versions (Requests is also
bumped to match to the respective minimum version enforce by Sphinx).
This will also allow the repository to drop a series of
compatibility-related implementation in following commits.

Signed-off-by: James Knight <james.d.knight@live.com>