build configuration-specific exceptions
Updating the configuration checks logic to raise specific configuration
exception cases instead of a generic configuration error. This is to
help avoid rendering content of any string literals in the stack trace
outside the message itself (EM101).

In addition, the exception messages built have been updated to take
advantage of f-strings (to avoid the need to format these messages), as
well as performing proper exception chaining.

Signed-off-by: James Knight <james.d.knight@live.com>