tests: provide a unique output path to support parallel tox runs
When populating an output directory for a unit test, build the path
taking into consideration the active tox environment name (if any). If
one is set, use the environment name as a sub-folder to place between
the output folder and a container/test-method name. This should prevent
any issues between various unit tests running, where each environment
is separated.

Signed-off-by: James Knight <james.d.knight@live.com>