config: capture non-str value error for enum check
The enumeration check for a string type can help ensure a provided
string matches to a specific enumeration type. However, if a user
defines a non-string value, the configuration check will fail with an
attribute error. To prevent this exception cause so we can manage our
own configuration error, also include `AttributeError` to the exception
list to capture.

Signed-off-by: James Knight <james.d.knight@live.com>