use non-private vars for navnode flags
When trying to coordinate navigational nodes on references between a
builder and translator, a `_navnode` prefix was used on these reference
nodes. This extension should avoid using private variables, which in
theory could be conflicting with any existing private variables that
could be used. To avoid this, instead use a Confluence-specific prefix
(`cbe_`; Confluence Builder extension) to have a more unique name for
nodes we do not control but lazily use to pass information.

Signed-off-by: James Knight <james.d.knight@live.com>