compat: drop assumption of sphinx modified inline-all-toctrees
Effort was never completed/pushed towards submitting tweaks to Sphinx's
`inline_all_toctrees` implementation to support a non-replacement
variable of this call. Trying to push this upstream is not a priority
at this time. For now, simplify the implementation to remove the todos
and suppressed assumption that the change was going to be made available
in a specific Sphinx version.

Signed-off-by: James Knight <james.d.knight@live.com>