ruff: update exceptions
Updating the ruff configuration to include all various linter rules to
ignore for this project. A series of linter issues have been resolved
over the past few days of maintenance work. Along with the exceptions
noted in this configuration (at this time, due to preference; and may
change over time), this puts running ruff into a passing state.

Signed-off-by: James Knight <james.d.knight@live.com>