publisher: adding more helpful unknown confluence instance exception
Provide a more helpful message on an initial 404 error message. If the
first request fails with a 404, aside from a bad URL, this could also
mean a bad/restricted space. Try to convey to the user that this
extension is not aware which specific reason the request failed, but
give more options to say to check a combination of the server URL, the
space ID and authentication options.

Signed-off-by: James Knight <james.d.knight@live.com>