publisher: add predefined rest client attribute
Adds a `rest` client attribute by default. This allows a user of this
class to invoke `disconnect` without having to have worried about first
invoking `connect`.

This change is primarily for debugging purposes, to avoid the chances of
throwing an attribute exception when a test's connect stage is unable to
be run (but a cleanup operation is invoked anyways).

Signed-off-by: James Knight <james.d.knight@live.com>