doc: tweak rtd badge icon for an improved alignment
It appears when the ReadTheDocs injected badge does not appear to align
the top-left book icon. Seems to be a consistent issue across a series
of browsers and have been observed for quite sometime. For now, adding
a tweak into this document's documentation stylesheet to provide a
better offset for now.

Signed-off-by: James Knight <james.d.knight@live.com>