rest: only close session if it was created
Provide a cleanup check to only close a session instance if a session
was created. This is to help select corner cases where a request to
`_setup_session` fails to create a session and throws an exception.
While an error case, we do not want an additional exception generated
when cleaning up from a broken scenario.

Signed-off-by: James Knight <james.d.knight@live.com>