builder: add workaround for v2 parenthesis anchor links
When a page is configured to be published with a v2 editor, check if
any anchor links contain a parenthesis value. If so, replace the
expected target value with a manipulated value which works around an
issue (JST-937603) where parenthesis will remove an anchor link.

This change also introduces an internal configuration entry
`confluence_adv_disable_confcloud_74698`, which disables this workaround
(if the situation arises where Confluence Cloud corrects this issue and
this applied workaround breaks user's documentation; a quick method for
users of this extension to deal with this until a next release).

See also: https://support.atlassian.com/requests/JST-937603/
Signed-off-by: James Knight <james.d.knight@live.com>