config: add auth value quote checks
Adding configuration checks that will throw a warning if a password or
token value is detected to be wrapped in quotes. This is to help inform
users who may be passing in password/token values through shell scripts,
environment variables or more; which may have a risk of being nested in
quotes by accident.

Signed-off-by: James Knight <james.d.knight@live.com>