publisher: rework get-base-page-id to use common get-page call
When a request to `get_base_page_id` results in using a page title to
find the base page, it did so using a custom REST call in the method.
However, a call to `get_page` would achieve the same results.
Refactoring the call to use `get_page` to simplify the implementation.

Signed-off-by: James Knight <james.d.knight@live.com>