util: strip api v2 url
When a URL is defined by a user, this extension will help strip off the
API sub-folder path commonly used by Confluence. This is since this
extension manages which API version to call (and also allows users to
override). While we strip the v1 API post-fix, we do not do the same if
an API v2 post-fix was added. Updating the logic to strip both v1 and v2
API path entries.

Signed-off-by: James Knight <james.d.knight@live.com>