allow registering entities for non-translating builders
This extension opted towards not registering directives, nodes and roles
into the Sphinx application in order to be flexible for users opting to
use directives/roles without `only` directive protection. This could
result in exceptions being raised when a translator (such as `html`)
could not processing a generated Confluence node type.

However, this also causes issues for unique builders such as Sphinx's
linkcheck builder or sphinxcontrib.spelling [1], where warnings are
generated for unknown directives/roles, making it hard to know if a
build is failing for either a bad-link/spelling versus the use of an
unknown directive/role.

In attempts to be even more flexible, this extension will permit the
registration for these entities when it detects that an active builder
has no translator (i.e. low risk of an unknown node exception). This
should prevent any undesired unknown warnings for builders who do not
explicitly care about unknown node types.

Finally, to support any odd case that has not been considered which may
block a user from registering these entities, this extension has added
an `confluence_adv_disable_init` option which will always restrict
registration for a documentation set.

[1]: https://sphinxcontrib-spelling.readthedocs.io/

Signed-off-by: James Knight <james.d.knight@live.com>