correct space key usage for generated search
The search macro generated for a search page should be using the
configured space key as the target space to search on. However, when
updates to this extension's configuration options were changes to
identify the option as a "key" over a "name", the search template was
not updated to use this value. This would result in the space value
being assigned to `None`; correcting.

Signed-off-by: James Knight <james.d.knight@live.com>