translator: simplify v2 autodocs renderings
The following commit simplifies the markup content generated for
autodocs related nodes. Attempts to gracefully format with various
workarounds (e.g. nesting with panels and layouts) seems to constantly
change as Confluence updates its ADF processing. There does not appear
to be a perfect way to layout autodocs content -- so we will no longer
try (that much). We will rely on panels for now (since it appears the
only way to do some level of block indentation of content without
borders/colors), and all other field entries will be just added into
simple paragraphs. This allows the content to be somewhat readable for
users and does not add a lot of tags that makes the viewing experience
slow on v2 editors.

We can look at this in the future if there are either alternative ways
to render autodocs, or get to a point where we may not even bother.

Signed-off-by: James Knight <james.d.knight@live.com>