publisher: rework page property processing
Calls to `get_page_property` should return `None` for non-entries on a
page. This was the original intent with the implementation, but final
tweaks for initial APIv2 support made the default result a partially
populated property instance. This commit cleans up this to restore the
original intent to have undefined properties return `None`, as well as
updates various locations to use a default property value for each
applicable section. Specifically:

- `CB_PROP_KEY` returns a prepared empty value dictionary, since the
   requests will push up a new property change
- Property fetching for v2 will return a `None` populated value to
   explicitly indicate that the requests should support this type of
   metadata, as well as that Confluence has reported no value exists
- Update events will populate with a fallback value `None`, which will
   be updated with the provided property value for the expected change
   to the property value.

Signed-off-by: James Knight <james.d.knight@live.com>