publisher: skip property version bumps for non-populated entries
This extension tracks certain property entries as an indication that
a property is supported for this instance (e.g. on Confluence Cloud, a
content width property exists, but not on Confluence Server (at this
time?)). However, even if we track support, it does not mean that
Confluence has reported the property has been defined for a page.

When we publish property updates for a page, we perform a version bump
to indicate to Confluence the property is being updated. Although, for
non-defined property entries, the value is set to `None` with no version
entries. This causes the implementation to fail since there is no
version value to bump. This commit corrects this by checking that the
previous value does exist before attempting any version bump; otherwise,
it treats the property update event as if it was created a new property
entry.

Signed-off-by: James Knight <james.d.knight@live.com>