flag file/link transforms as deprecated
The advanced overrides for tailoring file/link transformations existed
since the start of this extension's life, but the options are not
entirely useful* for end users. It is unknown if anyone has ever needed
to override these options. In addition, setting them make a project's
configuration unpicklable.

For now, will mark this options as deprecated. This will give time for
users to report if these options are desired before the configuration
options are removed from this extension.

Signed-off-by: James Knight <james.d.knight@live.com>