translator: support newer sphinxcontrib-video sources
Updating the translator to support sphinxcontrib-video's newer source
tracking in a node (as of v0.2.1; [1]). For now, we will grab the first
entry in the sources list as the source to use. This can be reworked
later if a use case arises.

See also: https://github.com/sphinx-contrib/video/commit/959aeff41240ddb75955137fda45c9cce9d0bb8f

Signed-off-by: James Knight <james.d.knight@live.com>