translator: avoid creating duplicate anchors
When building title anchors, the title node will look at assigned name
references to build anchors. Before building an anchor, the target will
be resolved for the case of special targets (e.g. references to headers
which use a Confluence-generated anchor target). In some scenarios, a
resolved reference may repeat the same target anchor to create. This
commit tweaks the process to avoid creating additional anchors of the
same name, if an anchor was already created for the title node.

Signed-off-by: James Knight <james.d.knight@live.com>