redact authorization header in debug bumps
When a user has configured logging for `headers`, this will print a
series of raw headers to the configured standard output stream. This is
primarily to help aid in debugging and development scenarios. To help
avoid situations where users may want to share the authorization header,
force redact the header's value when printing.

This commit also introduces the option `headers_raw` to allow advanced
users to dump the raw header value with no redaction, if ever needed.

Signed-off-by: James Knight <james.d.knight@live.com>