doc: improve styled configuration references
With reStructuredText/Sphinx, there does not appear to be a way to take
advantage of the `:ref:` role and apply a literal style processing on
select elements. This extension's documentation would handle this by
first adding a literal description of an element, followed by a reference
named "ref" -- not ideal.

This change adds a custom domain/role to support a `lref` role, the
"literal-styled reference" role. For generated references using this
role, they will be manipulated/forwarded to Sphinx's standard domain to
handle as a normal cross reference. When resolved, the custom domain
will then wrap the inner node with a literal node. This allows both
having a page-to-page reference that is also styled as a literal.

Signed-off-by: James Knight <james.d.knight@live.com>