cleanup publish debug options (format/presentation)
Updating the configuration handling of publish debugging options to
promote the uses of dashes of underscores (e.g. `headers-and-data` over
`headers_and_data`). Both values are still accepted, but the dash
variant will be the value shown in documentation.

This commit also adjusts the configuration check to explicitly list
supported options. The previous use of populating values from the
enumeration does not work for all desired entries. Best to list the
explicit values we want to inform a user on.

Signed-off-by: James Knight <james.d.knight@live.com>