config: configure notifications to default off
As this extension has evolved over the years, the option
`confluence_disable_notifications` was introduced to help suppress
multiple notifications generated from publishing events. At the time of
introduction, the feature was not enabled by default since it did
result in changing page requests made and there was not an immediate
desire to change the default state of things.

Consider the common application of this extension, most users may not
expect notifications to be generated for automated publish events. To
make things easier for installations, we are changing the default mode
of this extension to suppress page notifications. Users who prefer to
have notification events generated are now recommended to explicitly
configure the option.

Signed-off-by: James Knight <james.d.knight@live.com>