translator: introduce tag decorators
Providing a decorator implementation in attempt to automatically append
tags when processing a type of node. This is an attempt to reduce the
complement logic between the visit and depart nodes, mainly when it
becomes a bit too complex to track what end-tags needs to be applied in
the depart method.

Applying to the container node for now. May be expanded on other nodes,
or removed if deemed to be not helpful.

Signed-off-by: James Knight <james.d.knight@live.com>