tests: provide explicit project name in postfix hash check
Generating hashes in postfixes relies on the project name. As of Sphinx
v7.4.x, the default project name has been changes (was `Python` but now
reports `Project name not set`). Updating the unit test to provide an
explicit project name to ensure testing is consistent between the
different versions of Sphinx tested against.

Signed-off-by: James Knight <james.d.knight@live.com>