adjust multiple warning level logs to informational
A series of retry events when publishing can report certain events as
warnings. The prospect Sphinx v8.1.x release reworks how warnings can
trigger a fail state and `skip_warningiserror` becomes deprecated with
these changes. This can cause issues for users who wish to run with
warnings-as-failure but may experience publishing issues that can be
recovered on during a run. To prevent issues, updating a series of
warning messages to be informational messages.

Signed-off-by: James Knight <james.d.knight@live.com>