tests: support sphinx's exception-on-warning support
Sphinx's development branch has updated it's implementation to no longer
generate exceptions on an immediate warning by default [1]. This is
desired for this extension's tests so the Sphinx application will enable
this capability.

In addition, the `skip_warningiserror` call is obsolete. With
exception-on-warning enabled, a series of tests need to be updated to
operate with a relaxed state; updating.

[1]: https://github.com/sphinx-doc/sphinx/pull/12743

Signed-off-by: James Knight <james.d.knight@live.com>