builder: correct title track database check
When processing doctree targets to track titles across multiple
documents, the check for the database would fail to function if the
database did not any initial entries. This would properly prevent
tracking conflicting title documents when running with a
`singleconfluence` builder. Adjusting the check to ignore/suppress if
the provided database entry is `None`.

Signed-off-by: James Knight <git@jdknight.me>