.github: move linting jobs into their own workflow
The ruff lint environment was put in a non-passed state since it was run
on a previous change set. This is since a change in `doc/conf.py`
resulted in a failure and a change on this file only triggered the
documentation workflow to run. An update to the workflow trigger events
is needed to ensure `doc/conf.py` trigger a lint check. Although, we do
not want to force an entire build if the documentation's configuration
has just changed. Moving linting into its own workflow handle these
scenarios.

Signed-off-by: James Knight <git@jdknight.me>