translator: replace spaces with dashes when resolving title anchors
When attempting to determine the anchor target for a named title entry,
replace any spaces found in a named entry to ensure it can map to an
appropriate anchor value (since anchors have no spaces). This should
ensure the sanity check for a target lookup help finds the title has a
valid target to ensure the anchor is created.

Signed-off-by: James Knight <git@jdknight.me>